[MOV] hr_holidays: move function from employee to employee base
The function `_get_consumed_leaves`, that was in the hr_employee model
is used in a compute method of the hr_employee_base model.
This was causing some issues as the inhertiance is from hr_employee_base
to hr_employee. This commit fixes the issue by moving the function
to hr_employee_base.

Steps to reproduce:
- Install the hr_holidays module.
- Create a user with no HR rights.
- Connect with that user and go to Employees.
- Click on an employee, activate debug mode and click on "View Raw Record Data"

task-4742768

Forward-port-of: odoo#207026
X-original-commit: 4ec23020aa10d8735cd15717de9a99b00a0e137c
